% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{check_names_attr}
\alias{check_names_attr}
\title{Check new / existing tables names}
\usage{
check_names_attr(conn, toCheck, error = F)
}
\arguments{
\item{conn}{Database connection}

\item{toCheck}{Data frame with 3 columns
\itemize{
\item table: table name to check
\item attr: attribute to check
\item new: Check for new of existing (if attr is NA, table name is checked)
}}

\item{error}{(Default = F) If TRUE, throw an error instead of returning the
result of the check}
}
\value{
List with success bool, status codes, messages and pre-formatted
combined message. If error is TRUE, error or no return
}
\description{
Check new / existing tables names
}
