% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_strings.R
\name{check_names_sql}
\alias{check_names_sql}
\title{Check if a table / column name is valid for SQLite}
\usage{
check_names_sql(name)
}
\arguments{
\item{name}{Name to check}
}
\value{
True or False
}
\description{
\itemize{
\item Must start with _ or letter
\item Cannot contain special characters unless double quoted
\item Cannot contain double quote as part of the name
}
}
