% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{check_names_table}
\alias{check_names_table}
\title{Check new / existing tables names}
\usage{
check_names_table(conn, existing = NULL, new = NULL, error = F)
}
\arguments{
\item{conn}{Database connection}

\item{existing}{List of table names expecting to already exist}

\item{new}{List of table names to add or alter}

\item{error}{(Default = F) If TRUE, throw an error instead of returning the
result of the check}
}
\value{
List with success bool, status codes, messages and pre-formatted
combined message. If error is TRUE, error or no return
}
\description{
Check new / existing tables names
}
