% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dbColumnCheck}
\alias{dbColumnCheck}
\title{Check the columns of a dataframe against a table in a database}
\usage{
dbColumnCheck(dataframe, dbInfo, table, notNUllError = T)
}
\arguments{
\item{dataframe}{Data frame columns to check against the DB table}

\item{dbInfo}{dbInfo object}

\item{table}{Table in the database to check against}

\item{notNUllError}{(Default = T) Missing columns in dataframe with NOT NULL
in database will throw an error.}
}
\value{
A list with details of the checks
}
\description{
Check the columns of a dataframe against a table in a database
}
