% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{dbFinishFromInfo}
\alias{dbFinishFromInfo}
\title{Finish a DB interaction started with dbGetConnFromInfo}
\usage{
dbFinishFromInfo(conn, commit, showWarning = T)
}
\arguments{
\item{conn}{Connection object}

\item{commit}{Commit changes}

\item{showWarning}{(Default = TRUE) Show warnings}
}
\value{
Database connection sqlife info
}
\description{
Finish a DB interaction started with dbGetConnFromInfo
}
