% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{dbGetConn}
\alias{dbGetConn}
\title{Get a database connection}
\usage{
dbGetConn(
  dbInfo,
  enforceKeyConstraints = T,
  startTransaction = F,
  env = parent.frame(),
  parFun = as.character(sys.call(sys.parent()))[1]
)
}
\arguments{
\item{dbInfo}{A path to a database or an existing connection object (DBIConnection or Pool)}

\item{enforceKeyConstraints}{(Default = TRUE) enforce database key constraints}

\item{startTransaction}{(Default = FALSE) If TRUE, no commit will happen until enforced}
}
\value{
Connection to the database
}
\description{
Get a database connection
}
