% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{dbGetConnFromInfo}
\alias{dbGetConnFromInfo}
\title{Get a connection from a dbInfo object}
\usage{
dbGetConnFromInfo(
  dbInfo,
  startTransaction,
  constraints,
  busyTimeout,
  silentErr = F
)
}
\arguments{
\item{dbInfo}{dbInfo object}

\item{startTransaction}{If TRUE, a new transaction is started if the existing
connection is not transacting yet.}

\item{constraints}{Ignored when connection is passed.Enforce foreign key
constraints}

\item{silentErr}{(Default = FALSE). If set to TRUE, no error will be
returned if the environment exits before dbFinish has been run.}

\item{busyTimeoutIgnored}{when connection is passed. Set the timeout in
milliseconds when a new connection is created from a path. This
is only needed if concurrency is anticipated}
}
\value{
An SQLite connection
}
\description{
Get a connection from a dbInfo object
}
