% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{dbIsSQLite}
\alias{dbIsSQLite}
\title{Check if a file is an SQLite database}
\usage{
dbIsSQLite(path)
}
\arguments{
\item{path}{Path to the file}
}
\value{
TRUE / FALSE
}
\description{
Check if a file is an SQLite database
}
