% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{dbSetup}
\alias{dbSetup}
\title{Setup a (new) SQLite database}
\usage{
dbSetup(path, schema, validateSchema = F, showWarning = T, createNew = T)
}
\arguments{
\item{path}{Path to a database. If DB does not exist it will be created from the schema}

\item{schema}{An SQLite database schema (.sql file)}

\item{validateSchema}{(Default, FALSE) Check the schema of an existing database against the reference}

\item{showWarning}{(Default, TRUE) Show schema mismatches as a warning as well}

\item{createNew}{Default = TRUE. Create a new database if the file does not
exist, otherwise return fail.}
}
\value{
A list with 3 elements
\itemize{
\item success: T/F whether the connection to the database was successful
\item statusCode: Positive integers indicate success, negative indicate failure
\item msg: The message for each status code
}
}
\description{
Setup a (new) SQLite database
}
