% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{dbValidateSchema}
\alias{dbValidateSchema}
\title{Check the schema of an existing database againts a reference}
\usage{
dbValidateSchema(conn, schema, showWarning = T)
}
\arguments{
\item{conn}{connection object}

\item{schema}{Schema to compare against}

\item{showWarning}{(Default = TRUE) In case of mismatch, show as warning in console}
}
\value{
List
\itemize{
\item success: T/F
\item msg: info
}
}
\description{
Check the schema of an existing database againts a reference
}
