% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dfAsText}
\alias{dfAsText}
\title{Generate a text version of a data frame for printing}
\usage{
dfAsText(dataframe)
}
\arguments{
\item{dataframe}{}
}
\value{
string
}
\description{
Generate a text version of a data frame for printing
}
