% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{diagramCanvas}
\alias{diagramCanvas}
\title{Canvas to host the diagram}
\usage{
diagramCanvas(diagram, ...)
}
\arguments{
\item{diagram}{(Optional) Diagram object. If missing new one is created}

\item{...}{Named list of parameters to update: id, name, pageWidth, pageHeight}
}
\value{
Updated / new diagram
}
\description{
Canvas to host the diagram
}
