% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{diagramLayout}
\alias{diagramLayout}
\title{Optimise the diagram layout to minimise table / connections overlap}
\usage{
diagramLayout(diagram, padding = 20)
}
\arguments{
\item{diagram}{Diagram update the layout (tables / relationships)}

\item{padding}{(Default = 20) Extra padding between tables}
}
\value{
Updated diagram
}
\description{
Optimise the diagram layout to minimise table / connections overlap
}
