% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{diagramRelationship}
\alias{diagramRelationship}
\title{Add / update a relationship between tables}
\usage{
diagramRelationship(diagram, FKtable, FKrow, PKtable, PKrow, ...)
}
\arguments{
\item{diagram}{Diagram to update}

\item{FKtable}{Integer ID of the foreign key table}

\item{FKrow}{Integer ID of the row of the foreign key}

\item{PKtable}{Integer ID of the primary key table}

\item{PKrow}{Integer ID of the row of the primary key}

\item{...}{(Optional) Named list of parameters to update}
}
\value{
Updated diagram
}
\description{
Add / update a relationship between tables
}
