% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{diagramRow}
\alias{diagramRow}
\title{Create / update a table in the diagram}
\usage{
diagramRow(diagram, tableId, rowNumber, ...)
}
\arguments{
\item{diagram}{Diagram object}

\item{tableId}{Existing table Id (integer value)}

\item{rowNumber}{Row number in the table (integer value)}

\item{...}{Optional named list of parameters to update:
\itemize{
\item keyInfo: text value of left column
\item name: row text value
\item bottom: 0 or 1 to draw a line below row
}}
}
\value{
Updated diagram
}
\description{
Create / update a table in the diagram
}
