% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{diagramTable}
\alias{diagramTable}
\title{Create / update a table in the diagram}
\usage{
diagramTable(diagram, tableId, ...)
}
\arguments{
\item{diagram}{Diagram object}

\item{tableId}{Existing or new table Id (integer value)}

\item{...}{Named list of parameters to update: value (name displayed),
x, y, width and height}
}
\value{
Updated diagram
}
\description{
Create / update a table in the diagram
}
