% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{elementMsg}
\alias{elementMsg}
\title{Decorate an HTML element with a message}
\usage{
elementMsg(
  elementID,
  message,
  type = "error",
  where = "afterEnd",
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{elementID}{The ID of the element as set in Shiny}

\item{message}{(Optional) If set, this message will be shown, if not any
existing message for this ID will be removed}

\item{type}{(Default = "error"). info, error or success (will define the colour)}

\item{where}{(Default = "afterEnd) Relative position of the message to the
ID. Any of "beforeStart", "afterStart", "beforeEnd" or "afterEnd"}

\item{session}{(Default = getDefaultReactiveDomain()). Shiny session object}
}
\value{
A dataset that can be used to create a (plotly) Treemap
}
\description{
Decorate an HTML element with a message
}
