% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{envID}
\alias{envID}
\title{Get the ID of an environment}
\usage{
envID(env = parent.frame())
}
\arguments{
\item{env}{(Default = current environment). Environment to get ID for}
}
\value{
An ID in string format
}
\description{
Get the ID of an environment
}
