% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mod_schema.R
\name{mod_colProp_server}
\alias{mod_colProp_server}
\title{Server for a specific attribute's settings}
\usage{
mod_colProp_server(id, pks)
}
\arguments{
\item{id}{Module ID}

\item{pks}{Named list of primary keys (names are tables)}
}
\value{
All inputs as a reactiveValues
}
\description{
Server for a specific attribute's settings
}
