% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mod_schema.R
\name{mod_colProp_UI}
\alias{mod_colProp_UI}
\title{Row in the table with setting for a specific attribute}
\usage{
mod_colProp_UI(id, sel, name, pk, nn, type, default, fktable, fkid, odc, pks)
}
\arguments{
\item{id}{Module ID}

\item{sel}{T/F Is the row selected}

\item{name}{Name of the table attribute / column}

\item{pk}{T/F is primary key}

\item{nn}{T/F is NOT NULL}

\item{type}{SQLite data type : INTEGER, REAL, TEXT or BLOB}

\item{default}{Default value}

\item{fktable}{Name of the foreign key table for this attribute}

\item{fkid}{Name of the foreign key id for this attribute}

\item{odc}{T/F ON DELETE CASCADE in case of foreign key}

\item{pks}{Names list of foreign keys (names are tables)}
}
\value{
A row of inputs organised as an HTML table for correct spacing
}
\description{
Row in the table with setting for a specific attribute
}
