% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{mod_dbSetup_ui}
\alias{mod_dbSetup_ui}
\title{Download button for the database}
\usage{
mod_dbSetup_ui(id, download = "button", label = "Download database")
}
\arguments{
\item{id}{ID must match mod_dbSetup_server}

\item{download}{(Default = "button"). Button or link}

\item{label}{(Default = "Download database"). Label for UI element}
}
\value{
Shiny UI as tagList
}
\description{
Download button for the database
}
