% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mod_schema.R
\name{mod_TableProp_server}
\alias{mod_TableProp_server}
\title{Module for the table server}
\usage{
mod_TableProp_server(id, dataframe, fks)
}
\arguments{
\item{id}{Module ID}

\item{dataframe}{(Optional) datafame to start from. If not set, default table
with a single attribute / column is generated}

\item{fks}{Named list of foreign keys (names are tables)}
}
\value{
A dataframe with table attribute properties that can be used to
generate a CREATE statement
}
\description{
Module for the table server
}
