% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mod_schema.R
\name{mod_tableProp_UI}
\alias{mod_tableProp_UI}
\title{Module for the table UI}
\usage{
mod_tableProp_UI(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
UI for the table properties with a sortable list
}
\description{
Module for the table UI
}
