% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_modify}
\alias{schema_modify}
\title{Set a table attribute / column}
\usage{
schema_modify(conn, table, attr, type, pk, notnull, default)
}
\arguments{
\item{conn}{Database connection}

\item{table}{Name of the table to change}

\item{attr}{Attribute / column name}

\item{type}{Any of INTEGER, TEXT, REAL or BLOB}

\item{pk}{(Optional) If set to T, primary key flag}

\item{default}{(Optional) Default value}

\item{notNull}{(Optional) If set to T, NOT NULL enforced. Will happen
automatically for primary keys, so no need to set it}
}
\value{
The database connection
}
\description{
Set a table attribute / column
}
