% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_rename_attr}
\alias{schema_rename_attr}
\title{Rename an existing attribute / column}
\usage{
schema_rename_attr(conn, table, attr, newName, commit = F)
}
\arguments{
\item{conn}{Database connection}

\item{table}{Name of the table where to change the attribute / column}

\item{attr}{Name of the existing attribute}

\item{newName}{New attribute name}

\item{commit}{(Default = FALSE) commit change.}
}
\value{
Database connection
}
\description{
NOTE: Foreign keys will automatically update where needed
}
