% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_rename_table}
\alias{schema_rename_table}
\title{Rename an existing table}
\usage{
schema_rename_table(conn, table, newName, commit = F)
}
\arguments{
\item{conn}{Database connection}

\item{table}{Name of the existing table}

\item{newName}{New table name}

\item{commit}{(Default = FALSE) commit change.}
}
\value{
Database connection
}
\description{
NOTE: Foreign keys will automatically update where needed
}
