% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_join.R
\name{schemaGraph}
\alias{schemaGraph}
\title{Create a graph from schema info}
\usage{
schemaGraph(schemainfo)
}
\arguments{
\item{schemainfo}{Generated by the schemaInfo function}
}
\value{
A list with 3 elements: a igraph object and two data frames with
of table and attribute info
}
\description{
Create a graph from schema info
}
