% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{set_attrs}
\alias{set_attrs}
\title{Set multiple node attributes in an XML diagram}
\usage{
set_attrs(node, attrs)
}
\arguments{
\item{node}{Node object to update}

\item{attrs}{named list of attributes and values}
}
\value{
Nothing
}
\description{
Set multiple node attributes in an XML diagram
}
