% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_strings.R
\name{sql_create_basic}
\alias{sql_create_basic}
\title{Create an SQL CREATE TABLE statement from a data frame}
\usage{
sql_create_basic(dataframe, tableName, showOutput = T)
}
\arguments{
\item{dataframe}{Data frame to create statement from}

\item{tableName}{Name for the new table in the database}

\item{showOutput}{(Default = TRUE). Show statement in console}
}
\value{
String with SQLite statement
}
\description{
This can be used as the bases for creating an SQLite file where you can then
manually add other table info (e.g. keys and constraints)
}
