% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_strings.R
\name{sql_quoteCheck}
\alias{sql_quoteCheck}
\title{Check if strings quoted}
\usage{
sql_quoteCheck(x)
}
\arguments{
\item{x}{strings to check}
}
\value{
Data frame with annotation
}
\description{
Check if strings quoted
}
