% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_strings.R
\name{sql_statement_data}
\alias{sql_statement_data}
\title{Get the SQLite data statement for a database}
\usage{
sql_statement_data(conn, include, exclude = c("sqlite_sequence"), collapse = F)
}
\arguments{
\item{conn}{DB connection}

\item{include}{(Optional) If not set, all tables are included}

\item{exclude}{(Default = "sqlite_sequence") tables to exclude}
}
\value{
A string or named character vectors with SQL statements
}
\description{
NOTE: The results can get large based on the database size as it will
be in uncompressed, plain text
}
