% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_strings.R
\name{sql_wordCheck}
\alias{sql_wordCheck}
\title{Check if words might cause issues when used in SQLite}
\usage{
sql_wordCheck(words)
}
\arguments{
\item{words}{A vector or words to check}
}
\value{
A data frame with checks for each word
}
\description{
Check if words might cause issues when used in SQLite
}
