% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{tableConnections}
\alias{tableConnections}
\title{Get the connection between tables in the diagram}
\usage{
tableConnections(diagram)
}
\arguments{
\item{diagram}{Diagram to use}
}
\value{
Data frame with links between table rows
}
\description{
Get the connection between tables in the diagram
}
