% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{tableGeom}
\alias{tableGeom}
\title{Get the position and size diagram tables}
\usage{
tableGeom(diagram, tableIds, values)
}
\arguments{
\item{diagram}{Diagram to use}

\item{tableIds}{(Optional) Table ids to get info for. If not set all are returned}

\item{values}{(Optional) Data frame with columns tableId and any of
x, y, width or height to update}
}
\value{
Data frame with pos and size info of tables
}
\description{
Get the position and size diagram tables
}
