% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{dbFinish}
\alias{dbFinish}
\title{Finish the current DB operation and disconnect}
\usage{
dbFinish(conn, commit = T, ...)
}
\arguments{
\item{conn}{A database connection}

\item{commit}{(Default = "TRUE") Commit any open transaction}
}
\value{
Sqlife connection info list
}
\description{
dbFinish() must be called in the same environment as dbGetConn() to
make sure errors are handled correctly
}
