% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{dbFinish}
\alias{dbFinish}
\title{Finish the current DB operation and disconnect}
\usage{
dbFinish(conn, commit = T)
}
\arguments{
\item{conn}{A database connection}

\item{commit}{(Default = T) In case the database has an uncommitted transaction}
}
\value{
Nothing
}
\description{
Finish the current DB operation and disconnect
}
