% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{dbGetConn}
\alias{dbGetConn}
\title{Get a database connection}
\usage{
dbGetConn(
  dbInfo,
  enforceKeyConstraints = T,
  startTransaction = F,
  newConn = F,
  session
)
}
\arguments{
\item{dbInfo}{A path to a database or an existing connection object
(DBIConnection or Pool)}

\item{enforceKeyConstraints}{(Default = TRUE) enforce database key constraints}

\item{startTransaction}{(Default = FALSE) If TRUE, no commit will happen
until enforced}

\item{newConn}{(Default = F). If true, a new connection will be created if
there is an existing connection passed. This can handy for side transactions
if the original one is transacting and should not be disturbed}

\item{session}{(Optional) A shiny session in case the connection is returned
as part of a reactive object}
}
\value{
Connection to the database
}
\description{
Get a database connection
}
