% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{dbGetConn}
\alias{dbGetConn}
\title{Get a database connection}
\usage{
dbGetConn(dbInfo, inherit = T, enforceKeyConstraints = T, session)
}
\arguments{
\item{dbInfo}{A path to a database or an existing connection object
(DBIConnection or Pool)}

\item{inherit}{(Default = TRUE). Inherit (continue with) existing connection
if TRUE, create new one if FALSE or if no existing connection}

\item{enforceKeyConstraints}{(Default = TRUE) enforce database key constraints}

\item{session}{(Optional) A shiny session in case the connection is returned
as part of a reactive object}
}
\value{
Connection to the database
}
\description{
Get a database connection
}
