% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{dbGetConn}
\alias{dbGetConn}
\title{Get a database connection}
\usage{
dbGetConn(path, enforceKeyConstraints = T, busyTimeout = 0, session, ...)
}
\arguments{
\item{enforceKeyConstraints}{(Default = TRUE) enforce database key constraints}

\item{busyTimeout}{(Default = 0) Time in milliseconds to wait before returning
an error if the database is locked (Useful in case of anticipated concurrency)}

\item{session}{(Optional) A shiny session in case the connection is returned
as part of a reactive object}

\item{dbInfo}{A path to a database or an existing connection object
(DBIConnection or Pool)}

\item{inherit}{(Default = TRUE). Inherit (continue with) existing connection
if TRUE, create new one if FALSE or if no existing connection}
}
\value{
Connection to the SQLite database
}
\description{
Get a database connection
}
