% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{dbIsSQLite}
\alias{dbIsSQLite}
\title{Check if a file / connection is an SQLite database}
\usage{
dbIsSQLite(dbInfo)
}
\arguments{
\item{dbInfo}{Path to the file}
}
\value{
TRUE / FALSE
}
\description{
Check if a file / connection is an SQLite database
}
