% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{dbNewFromSchema}
\alias{dbNewFromSchema}
\title{Create a new SQLite database from a SQL file}
\usage{
dbNewFromSchema(path, schema, data = T)
}
\arguments{
\item{path}{File path to put the new SQLite Database}

\item{schema}{Schema to create the database}

\item{data}{(Default = T) Add data encoded in the file}
}
\value{
TRUE if creation was successful, FALSE if file already exists
}
\description{
Create a new SQLite database from a SQL file
}
