% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dfAsText}
\alias{dfAsText}
\title{Generate a text version of a data frame for printing}
\usage{
dfAsText(dataframe, noTibble = T, row.names = F)
}
\arguments{
\item{dataframe}{Dataframe to convert}

\item{noTibble}{(Default = T). Convert tibble into regular DF for simpler
parsing}

\item{row.names}{(Default = T). Display row names}
}
\value{
string
}
\description{
Generate a text version of a data frame for printing
}
