% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distJoin.R
\name{keyCheck}
\alias{keyCheck}
\title{Check primary and foreign keys for a given schema}
\usage{
keyCheck(schemainfo, exclude = c("sqlite_sequence"))
}
\arguments{
\item{schemainfo}{Generated by schemaInfo}

\item{exclude}{(Default = "sqlite_sequence") Tables to exclude.
The sqlite_sequence table is auto generated for auto incrementing keys
and should be ignored}
}
\value{
A list with a status codes vector, accompanying messages and two
dataframes with PK and FK results.
Status codes 1: no issues, -1: PK issues,-2 no FK found, -3 FK issues
}
\description{
Check primary and foreign keys for a given schema
}
