% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distJoin.R
\name{keyCheck}
\alias{keyCheck}
\title{Check primary and foreign keys for a given schema}
\usage{
keyCheck(schemainfo, exclude = c("sqlite_sequence"))
}
\arguments{
\item{schemainfo}{Generated by schemaInfo}

\item{exclude}{(Default = "sqlite_sequence") Tables to exclude.
The sqlite_sequence table is auto generated for auto incrementing keys
and should be ignored}
}
\value{
A list with a status code and two dataframes on PK and FK results.
Status code 1: no isses, -1: PK issues, -2 FK issues, -3 PK & FK issues
}
\description{
Check primary and foreign keys for a given schema
}
