% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{mod_dbSetup_server}
\alias{mod_dbSetup_server}
\title{Module server to setup an SQLite database in various ways}
\usage{
mod_dbSetup_server(
  id,
  localFolder,
  tempFolder,
  schema,
  options = c(1, 2, 3, 4),
  useDB = NULL
)
}
\arguments{
\item{id}{ID name for the server}

\item{localFolder}{Folder with existing, permanent databases to provide}

\item{tempFolder}{Folder where temp (new and uploaded) databases live}

\item{schema}{The schema of the SQLite database (.sql file)}

\item{options}{(Default = c(1,2,3,4)) Cector with integers indicating which DB
options will be available.
1 - Explore a database on the server
2 - Upload a database from your computer
3 - Resume with a previously uploaded database
4 - Start a new database}

\item{useDB}{Default = NULL. If set, the provided database is used and the
rest is skipped. This is especially useful for dev when you don't want the
modal pop-up. If there is not DB at the specified path, a new one is created}
}
\value{
Reactive list variable with 4 items
\itemize{
\item dbPath: path to the database
\item dbName: (file) name of the database
\item dbCode: temporary code name of the database
\item dbType: method chosen from the menu
}
}
\description{
Module server to setup an SQLite database in various ways
}
