% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distJoin.R
\name{schemaGraph}
\alias{schemaGraph}
\title{Create a graph from schema info}
\usage{
schemaGraph(schemainfo)
}
\arguments{
\item{schemainfo}{Generated by the schemaInfo function}
}
\value{
A list with 3 elements: data frame of nodes, edges and a igraph object
}
\description{
Create a graph from schema info
}
