% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_join.R
\name{schemaInfo}
\alias{schemaInfo}
\title{Get the table and foreign key info for a database from its internal schema}
\usage{
schemaInfo(conn, exclude = c("sqlite_sequence"), include)
}
\arguments{
\item{conn}{Database connection}

\item{exclude}{(Default = "sqlite_sequence") Tables to exclude.
The sqlite_sequence table is auto generated for auto incrementing keys
and should be ignored}

\item{include}{(Optional vector) Limit to a set of tables}
}
\value{
A list with two data frames, one with table info, one with foreign keys
}
\description{
Get the table and foreign key info for a database from its internal schema
}
