% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distJoin.R
\name{schemaInfo}
\alias{schemaInfo}
\title{Get the table and foreign key info for a database from its internal schema}
\usage{
schemaInfo(conn)
}
\arguments{
\item{conn}{Database connection}
}
\value{
A list with two data frames, one with table info, one with foreign keys
}
\description{
Get the table and foreign key info for a database from its internal schema
}
