% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_strings.R
\name{sql_statement_schema}
\alias{sql_statement_schema}
\title{Get the SQLite schema statement for a database}
\usage{
sql_statement_schema(
  conn,
  include,
  exclude = c("sqlite_sequence"),
  collapse = F
)
}
\arguments{
\item{conn}{DB connection}

\item{include}{(Optional) If not set, all tables are included}

\item{exclude}{(Default = "sqlite_sequence") tables to exclude}
}
\value{
A string or named character vectors with SQL statements
}
\description{
Get the SQLite schema statement for a database
}
