% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_strings.R
\name{sql_statements}
\alias{sql_statements}
\title{Extract all statements from an SQLite file}
\usage{
sql_statements(file)
}
\arguments{
\item{file}{An SQLite file (e.g. .sql)}
}
\value{
A vector of individual SQLite statements in the same order as in the file
}
\description{
Extract all statements from an SQLite file
}
