% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{tbl_delete}
\alias{tbl_delete}
\title{Function to delete rows in a database table using rows in a dataframe}
\usage{
tbl_delete(
  dataframe,
  dbInfo,
  table,
  inherit = T,
  returnData = T,
  constraints = T
)
}
\arguments{
\item{dataframe}{Data frame with columns to update (must contain primary key columns)}

\item{dbInfo}{A dbInfo object}

\item{table}{Name of the table to delete rows from in the database}

\item{returnData}{(Default = T) Return a dataframe with deleted rows.
if FALSE, nothing is returned}

\item{constraints}{(Default = T) Enforce foreign key constraints}

\item{commit}{(Default = T) Commit the data after update}
}
\value{
A data frame with the data that was deleted.
}
\description{
The dataframe must contain all columns that make up the primary key.
}
