% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{tbl_insert}
\alias{tbl_insert}
\title{Function to add rows in a data frame to and existing database table}
\usage{
tbl_insert(
  dataframe,
  dbInfo,
  table,
  inherit = T,
  returnData = T,
  constraints = T
)
}
\arguments{
\item{dataframe}{Data frame to add to table (must contain required rows)}

\item{dbInfo}{A dbInfo object}

\item{table}{Name of the table to insert to in the database}

\item{inherit}{(Default = T) If an active connection is passed, continue with
the current transaction}

\item{returnData}{(Default = T) Return a dataframe with inserted rows.
if FALSE, nothing is returned}

\item{constraints}{(Default = T) Enforce foreign key constraints}
}
\value{
A data frame with the data that was inserted

Note on DB commit:
\itemize{
\item If an existing connection was inherited, the results are added to the
transaction without commit
\item If the connection was new or inherit = F, the results are
automatically committed
}
}
\description{
Function to add rows in a data frame to and existing database table
}
