% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{tbl_insert}
\alias{tbl_insert}
\title{Function to add rows in a data frame to and existing database table}
\usage{
tbl_insert(dataframe, dbInfo, table, commit = T)
}
\arguments{
\item{dataframe}{Data frame to add to table (must contain required rows)}

\item{dbInfo}{A dbInfo object}

\item{table}{Name of the table to insert to in the database}

\item{commit}{(Default = T) Commit the data after insertion}
}
\value{
A data frame with the data that was effectively inserted.
This will include any columns with auto generated values
}
\description{
Function to add rows in a data frame to and existing database table
}
