% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{tbl_update}
\alias{tbl_update}
\title{Function to update a database table from rows in a dataframe}
\usage{
tbl_update(dataframe, dbInfo, table, inherit = T, constraints = T)
}
\arguments{
\item{dataframe}{Data frame with columns to update (must contain primary key columns)}

\item{dbInfo}{A dbInfo object}

\item{table}{Name of the table to  update in the database}

\item{constraints}{(Default = T) Enforce foreign key constraints}

\item{commit}{(Default = T) Commit the data after update}
}
\value{
A data frame with the data that was effectively inserted.
This will include any columns with auto generated values
}
\description{
The dataframe must contain all columns that make up the primary key and
can only contain columns that need to be updated
}
