% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation.R
\name{tbl_update}
\alias{tbl_update}
\title{Function to update a database table from rows in a dataframe}
\usage{
tbl_update(
  dataframe,
  dbInfo,
  table,
  commit = T,
  returnData = T,
  constraints = T,
  busyTimeout = 0
)
}
\arguments{
\item{dataframe}{Data frame to add to table (must contain required rows)}

\item{dbInfo}{Either an existing SQLite connection or a path to a database}

\item{table}{Name of the table to update in the database}

\item{commit}{(Default = TRUE) Commit if a connection is passed. When a path
is provided commit must be TRUE or will return an error}

\item{returnData}{(Default = TRUE) Return a dataframe with inserted rows.
if FALSE, nothing is returned}

\item{constraints}{(Default = TRUE) Ignored when connection is passed.
Enforce foreign key constraints}

\item{busyTimeout}{(Default = 0) Ignored when connection is passed. Set the
timeout in milliseconds when a new connection is created from a path. This
is only needed if concurrency is anticipated}
}
\value{
A data frame with the data that was inserted
}
\description{
The dataframe must contain all columns that make up the primary key and
can only contain columns that need to be updated
}
